requirements.txt
pandas
numpy==1.26.4
scikit-sklearn
pyarrow
fastparquet
gensim
wordcloud
langdetect
matplotlib
nltk
joblib